Create `_types` file to hold protocol classes
